[decomp] Add test tracking core ATen operators
This adds an expect-test that finds the set of core ATen operators by
subtracting the operators with decomposition in core_aten_decompositions from the
set of all operators that have decompositions and could be decomposed.

This is useful because if you add a new decomposition but forget to add it to
the list of core decompositions, it will appear in the PR diff.

Also, by going through this list I have identified some operators where the
functional variant is decomposed, but not the inplace variant which must be an
oversight.

ghstack-source-id: d5f86eb2ab4b9b91a9f4fe047deff1617650f64b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/104262